requirements.txt
Flask==2.1.0
mysql-connector-python==8.0.28
azure-ai-textanalytics==5.1.0
